pyproject.toml
[tool.poetry]
name = "grip3"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{ include = "com", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10.0"
requests = "^2.25.1"
wheel = "^0.37.0"
pandas = "^2.2.2"
vantage6-algorithm-tools = "4.2.2"
vantage6-client = "4.2.2"
pyarrow = "^12.0.1"
fastparquet = "^2023.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
